new git 1399-6-12 Edit AuthUserController

+++ src/main/java/com/mahdi/sandogh/model/user/UserPrinciple.java
@@ -28,7 +28,7 @@ public class UserPrinciple implements UserDetails {
 
     private static final long serialVersionUID = 1L;
 
-    private Long id;
+    private Integer id;
     private String name;
     private String username;
     private String nationalId;
@@ -39,7 +39,7 @@ public class UserPrinciple implements UserDetails {
 
     private Collection<? extends GrantedAuthority> authorities;
 
-    public UserPrinciple(Long id, String name, String username, String nationalId, Boolean active, String password, Collection<? extends GrantedAuthority> authorities) {
+    public UserPrinciple(Integer id, String name, String username, String nationalId, Boolean active, String password, Collection<? extends GrantedAuthority> authorities) {
         this.id = id;
         this.name = name;
         this.username = username;
